pyproject.toml
[project]
name = "image-gen"
version = "0.1.0"
description = "CLI tool for SDXL image generation with safetensors models, LoRA and embedding support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diffusers>=0.27.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "safetensors>=0.4.0",
    "typer>=0.12.0",
    "Pillow>=10.0.0",
    "accelerate>=0.28.0",
    "omegaconf>=2.3.0",
    "peft>=0.10.0",
    "compel>=2.0.0",
    "piexif>=1.1.3",
]

[project.scripts]
image-gen = "image_gen.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/image_gen"]

[tool.hatch.build.targets.sdist]
include = ["src/image_gen"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
